+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * [Added] FileName column for macOS (Process file name)
 * [Added] `--thread` support for Windows, macOS and FreeBSD
 * [Added] External pager support for Windows by `[pager] command` (built-in pager is used if it is not set)
-* [Added] `show_other_users` of `[display]` section to show the processes of other users (false: only the current effective user)
+* [Added] `show_user_only` of `[display]` section and `--user` (`-u`) option to show the processes of one user (a name, a uid, or a SID on Windows, as well as `"myself"` and `"all"`)
 * [Changed] Group and Gid columns for Windows read the process token only when the column is displayed
 * [Fixed] System processes no longer being dropped silently on macOS
 * [Fixed] ReadBytes / WriteBytes divided by a mis-scaled interval (seconds added to milliseconds)